Implement get_directory_entry()
This will give directory entries each time it is called, until -1 is returned. It takes a cluster number as a pointer. This pointer will be modified the next time there is a cluster change. It also takes another pointer, offset, that likewise will be change, accumulating the offset of a directory entry (32 bytes), until there is a cluster change,  in wich case it will be set to 0.

Change the parameter of read_cluster() to take a cluster number instead of an offset