Fix set_label
The previous behaviour was to just write/override the first directory entry of the root directory, wich in most cases could be dangerous. And also:

* Implemented allocate_directory: this function created a new directory entry in the specified cluster (directory)

* Implemented write_to_fsi: this writes into the filesystem info structure and its backup

Implemented write_to_cluster: this writes to a cluster given a cluster number. Still needs to be worked on

Implemented read_fsi: this updates the information of the filesystem structure variables after a directory is added

Moved validate_83_name from a static context to a pulibc one